do not process if no picture has been taken

+++ main.sh
@@ -11,18 +11,23 @@ while :; do
 	else
 		fswebcam -r 800x600 --png -1 $curr
 	fi
-	if [ -z $previous ]
+	pictureTaken=$?
+
+	if [ $pictureTaken -eq 0 ]
 	then
-		previous=$curr
-	else
-		python pic.py "$previous" "$curr"
-		different=$?
-		if [ "$different" -eq 0 ]
+		if [ -z $previous ]
 		then
-			rm $curr
-		else
-			echo "Images different, keep $curr"
 			previous=$curr
+		else
+			python pic.py "$previous" "$curr"
+			different=$?
+			if [ "$different" -eq 0 ]
+			then
+				rm $curr
+			else
+				echo "Images different, keep $curr"
+				previous=$curr
+			fi
 		fi
 	fi
 	sleep 5
